pyproject.toml
[project]
name = "nonebot-plugin-fishing2"
version = "0.0.4"
description = "更好的电子钓鱼"
authors = [
    {name = "ALittleBot", email = "160833462+C14H22O@users.noreply.github.com"},
    {name = "Polaris_Light", email = "995905922@qq.com" }
]
dependencies = [
    "nonebot2>=2.2.1",
    "nonebot-adapter-onebot>=2.0.0-beta.1",
    "pydantic>=1.10",
    "nonebot-plugin-localstore>=0.6.0",
    "sqlalchemy>=2.0.27",
    "aiosqlite>=0.20.0",
    "nonebot-plugin-orm>=0.7.1",
]
readme = "README.md"
license = "MIT"
license-files = ["LICEN[CS]E*", "vendored/licenses/*.txt", "AUTHORS.md"]
requires-python = ">=3.8"
keywords = ["fishing"]
classifiers = []

[project.urls]
Homepage = "https://github.com/FDCraft/nonebot-plugin-fishing2"

[tool.poetry.dependencies]
nonebot2 = ">=2.2.1"
nonebot-adapter-onebot = ">=2.0.0-beta.1"
pydantic = ">=1.10"
nonebot-plugin-localstore = ">=0.6.0"
sqlalchemy = ">=2.0.27"
aiosqlite = ">=0.20.0"
nonebot-plugin-orm = ">=0.7.1"
